pyproject.toml
# Copyright 2022 Canonical Ltd.
# See LICENSE file for licensing details.

[project]
name = "rabbitmq-k8s"
version = "3.12"
requires-python = "~=3.12.0"

dependencies = [
    "cosl>=0.0.53",
    "ops",
    "pika",
    "pwgen",
    "rabbitmq-admin",
    "tenacity",
    "lightkube",
    "lightkube-models",
    "lightkube-extensions @ git+https://github.com/canonical/lightkube-extensions.git@main",
    "jsonschema",
]

[project.optional-dependencies]
dev = [
    "coverage",
    "flake8",
    "rabbitmq_admin",
    "ops[testing]",
    "pytest",
    "pytest-cov",
    "pyyaml",
    "jubilant",
]

# Testing tools configuration
[tool.coverage.run]
branch = true
concurrency = ["multiprocessing"]
parallel = true
source = ["."]
omit = [
    ".tox/*",
    "tests/*",
    "src/templates/*",
    "lib/charms/grafana_k8s/*",
    "lib/charms/loki_k8s/*",
    "lib/charms/prometheus_k8s/*",
    "lib/charms/traefik_k8s/*",
]

[tool.coverage.report]
show_missing = true

[tool.pytest.ini_options]
minversion = "6.0"
log_cli_level = "INFO"

# Formatting tools configuration
[tool.black]
line-length = 79

[tool.isort]
profile = "black"
multi_line_output = 3
force_grid_wrap = true

# Linting tools configuration
[tool.flake8]
max-line-length = 79
max-doc-length = 99
max-complexity = 10
exclude = [".git", "__pycache__", ".tox", "build", "dist", "*.egg_info", "venv"]
select = ["E", "W", "F", "C", "N", "R", "D", "H"]
# Ignore W503, E501 because using black creates errors with this
# Ignore D107 Missing docstring in __init__
ignore = ["W503", "E501", "D107", "E402", "E226"]
per-file-ignores = []
docstring-convention = "google"
# Check for properly formatted copyright header in each file
copyright-check = "True"
copyright-author = "Canonical Ltd."
copyright-regexp = "Copyright\\s\\d{4}([-,]\\d{4})*\\s+%(author)s"
